restructuring to a util based layout
