+++ orix/tests/test_s2grid.py
@@ -1,7 +1,7 @@
 import pytest
 
-from orix.grid.s1grid import S1Grid
-from orix.grid.s2grid import S2Grid
+from orix.utils.grid_utils.s1grid import S1Grid
+from orix.utils.grid_utils.s2grid import S2Grid
 
 
 @pytest.fixture
