+++ orix/utils/grid_utils/s2grid.py
@@ -18,7 +18,7 @@
 
 import numpy as np
 from orix.vector import Vector3d
-from orix.grid.s1grid import S1Grid
+from orix.utils.grid_utils.s1grid import S1Grid
 
 
 class S2Grid:
